Edited build.py

+++ c-unity/build.py
@@ -4,10 +4,15 @@ assert len(sys.argv) > 1
 assert sys.argv[1].endswith(".c")
 name = sys.argv[1][:-2]
 
-os.system("md build")
-os.system("g++ -c " + name + ".c -o build/" + name + ".o")
-os.system("g++ -c unity.c -o build/unity.o")
-os.system("ar rcs libs/libunity.a build/unity.o")
-os.system("g++ build/" + name + ".o -Llibs -lunity -o " + name + ".exe")
-os.system("rmdir /s /q build")
-os.system(name + ".exe")
+cmds = [
+    "md build", "g++ -c " + name + ".c -o build/" + name + ".o",
+    "g++ -c unity.c -o build/unity.o", "g++ -c unity.c -o build/unity.o",
+    "ar rcs libs/libunity.a build/unity.o",
+    "g++ build/" + name + ".o -Llibs -lunity -o " + name + ".exe", name + ".exe"
+]
+
+for cmd in cmds:
+    result = os.system(cmd)
+    if result: break
+
+if os.path.isdir("build"): os.system("rmdir /s /q build")
